+++ migrations/2024_12_10_000000_set_new_setting_for_discussions_replies_display_position.php
@@ -15,4 +15,5 @@ $allowedTags = [];
 
 return Migration::addSettings([
     'walsgit_discussion_cards_showRepliesOnRight' => 0,
+    'walsgit_discussion_cards_showLastPostInfo' => 0,
 ]);